Добавил select2, поправил баги

+++ app/Services/AdminPanel/Reviews/ReviewsService.php
@@ -7,13 +7,15 @@ use App\Http\Resources\AdminPanel\Review\ReviewsResource;
 use App\Models\AdminPanel\Review\Review;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\DB;
 
 class ReviewsService
 {
     public function createReview($data): JsonResponse|RedirectResponse
     {
         try {
-            $review = Review::create($data);
+            $review = Review::create(array_merge($data, ['writer_id' => Auth::user()->id]));
 
             if ($review) {
                 return redirect()->route('showReview', ['review' => $review->id])
@@ -50,22 +52,26 @@ class ReviewsService
     }
 
 
-    public function deleteReview($review): ReviewsResource|JsonResponse
+    public function deleteReview($review): array
     {
         try {
-            $deletedModel = $review;
+            if(Db::table('main_page_reviews')->where('review_id', $review->id)->exists()) {
+                return [
+                    'error' => 'Невозможно удалить обзор, он расположен на главной странице сайта.',
+                    'status' => 400,
+                ];
+            }
             $review->delete();
 
-            return (new ReviewsResource($deletedModel))
-                ->additional([
-                    'message' => 'Вы успешно удалили обзор! Если вы хотите его восстановить, то воспользуйтесь корзиной.',
-                    'status' => 200,
-                ]);
+            return [
+                'message' => 'Вы успешно удалили обзор!',
+                'status' => 200,
+            ];
         } catch (\Throwable $e) {
-            return response()->json([
-                'error' => 'Произошла ошибка при удалении обзора, пожалуйста, обратитесь к разработчику, Спасибо! Текст ошибки: ' . $e,
+            return [
+                'error' => 'Произошла ошибка при удалении обзора, пожалуйста, обратитесь к разработчику, Спасибо! Текст ошибки: ' . $e->getMessage(),
                 'status' => 400,
-            ], 400);
+            ];
         }
     }
